Adding in isDebug mode from Josh Robb to allow for easier output of mapping files in ActiveRecord.
git-svn-id: http://svn.castleproject.org:8080/svn/castle/trunk@1796 73e77b4c-caa6-f847-a29a-24ab75ae54b6

+++ ActiveRecord/Castle.ActiveRecord/Framework/Internal/Visitors/XmlGenerationVisitor.cs
@@ -47,22 +47,23 @@ namespace Castle.ActiveRecord.Framework.Internal
 			Dedent();
 			EndMappingNode();
 
-#if OUTPUTXML
-			String file = model.Type.Name + ".mapping.xml";
-
-			System.IO.File.Delete(file);
-
-			using(System.IO.FileStream fs = System.IO.File.OpenWrite(file))
+			if (ActiveRecordModel.isDebug) 
 			{
-				String xml = Xml;
+				String file = model.Type.Name + ".mapping.xml";
 
-				byte[] ba = System.Text.ASCIIEncoding.UTF8.GetBytes(xml);
+				System.IO.File.Delete(file);
 
-				fs.Write(ba, 0, ba.Length);
+				using(System.IO.FileStream fs = System.IO.File.OpenWrite(file))
+				{
+					String xml = Xml;
 
-				fs.Flush();
+					byte[] ba = System.Text.ASCIIEncoding.UTF8.GetBytes(xml);
+
+					fs.Write(ba, 0, ba.Length);
+
+					fs.Flush();
+				}
 			}
-#endif
 		}
 
 		public override void VisitModel(ActiveRecordModel model)
